predefined hull to be inserted
